feat: add simple params check

+++ src/movies/movies.controller.js
@@ -6,6 +6,14 @@ const router = express.Router();
 
 // Create a new movie
 router.post("/", AuthService.JwtMiddleware, (req, res) => {
+  const { title, description, genre, rating, year } = req.body;
+
+  if (!title || !description || !genre || !rating || !year) {
+    return res
+      .status(400)
+      .send("Missing title, description, genre, rating or year");
+  }
+
   const movie = Movies.createNewMovie(req.body);
   res.status(201).json(movie).send();
 });
@@ -17,18 +25,32 @@ router.get("/", AuthService.JwtMiddleware, (req, res) => {
 });
 
 router.delete("/:id", AuthService.JwtMiddleware, (req, res) => {
+  const id = req.params.id;
+  if (!id) {
+    return res.status(400).send("Missing id");
+  }
+
   MoviesService.deleteMovie(req.params.id);
   res.status(200).send();
 });
 
 router.get("/:title", AuthService.JwtMiddleware, (req, res) => {
+  const title = req.params.title;
+  if (!title) {
+    return res.status(400).send("Missing title");
+  }
+
   MoviesService.getMovieFromTitle(req.params.title)
     .then((movies) => res.status(200).json(movies))
     .catch((err) => res.status(500).json(err));
 });
 
 router.get("/date/:date", AuthService.JwtMiddleware, (req, res) => {
-  const date = req.params.date.replaceAll('-', '/');
+  const date = req.params.date.replaceAll("-", "/");
+  if (!date) {
+    return res.status(400).send("Missing date");
+  }
+
   MoviesService.getMovieFromDate(date)
     .then((movies) => res.status(200).json(movies))
     .catch((err) => res.status(500).json(err));
